Autowiring::completeArguments() moved to Resolver::autowireArguments()
